refactor(types): toggle-group
